Move files to create config.webhook.cert-manager.io
