+++ internal/apis/config/webhook/register.go
@@ -17,10 +17,9 @@ limitations under the License.
 package config
 
 import (
+	"github.com/jetstack/cert-manager/pkg/apis/config/webhook"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
-
-	"github.com/jetstack/cert-manager/pkg/apis/config"
 )
 
 var (
@@ -29,7 +28,7 @@ var (
 )
 
 // SchemeGroupVersion is group version used to register these objects
-var SchemeGroupVersion = schema.GroupVersion{Group: config.GroupName, Version: runtime.APIVersionInternal}
+var SchemeGroupVersion = schema.GroupVersion{Group: webhook.GroupName, Version: runtime.APIVersionInternal}
 
 // Resource takes an unqualified resource and returns a Group qualified GroupResource
 func Resource(resource string) schema.GroupResource {
