+++ internal/apis/config/webhook/v1alpha1/register.go
@@ -17,14 +17,13 @@ limitations under the License.
 package v1alpha1
 
 import (
+	"github.com/jetstack/cert-manager/pkg/apis/config/webhook"
+	"github.com/jetstack/cert-manager/pkg/apis/config/webhook/v1alpha1"
 	"k8s.io/apimachinery/pkg/runtime/schema"
-
-	"github.com/jetstack/cert-manager/pkg/apis/config"
-	"github.com/jetstack/cert-manager/pkg/apis/config/v1alpha1"
 )
 
 // SchemeGroupVersion is group version used to register these objects
-var SchemeGroupVersion = schema.GroupVersion{Group: config.GroupName, Version: "v1alpha1"}
+var SchemeGroupVersion = schema.GroupVersion{Group: webhook.GroupName, Version: "v1alpha1"}
 
 // Resource takes an unqualified resource and returns a Group qualified GroupResource
 func Resource(resource string) schema.GroupResource {
