+++ internal/apis/config/webhook/v1alpha1/zz_generated.conversion.go
@@ -22,10 +22,10 @@ limitations under the License.
 package v1alpha1
 
 import (
+	"github.com/jetstack/cert-manager/pkg/apis/config/webhook/v1alpha1"
 	unsafe "unsafe"
 
 	config "github.com/jetstack/cert-manager/internal/apis/config"
-	v1alpha1 "github.com/jetstack/cert-manager/pkg/apis/config/v1alpha1"
 	conversion "k8s.io/apimachinery/pkg/conversion"
 	runtime "k8s.io/apimachinery/pkg/runtime"
 )
