+++ internal/apis/config/webhook/v1alpha1/zz_generated.defaults.go
@@ -22,7 +22,7 @@ limitations under the License.
 package v1alpha1
 
 import (
-	v1alpha1 "github.com/jetstack/cert-manager/pkg/apis/config/v1alpha1"
+	"github.com/jetstack/cert-manager/pkg/apis/config/webhook/v1alpha1"
 	runtime "k8s.io/apimachinery/pkg/runtime"
 )
